requirements.txt
uvicorn==0.22.0
fastapi==0.97.0
asyncpg==0.27.0
coverage==7.2.7
greenlet==2.0.2
pytest==7.3.2
pytest-asyncio==0.21.0
httpx==0.24.1
sqlalchemy==2.0.16
pycryptodome==3.18.0
pytest-cov==4.1.0
psycopg2==2.9.6
alembic==1.11.1
pydantic==1.10.9